Login page, Home page style, Header style

+++ src/pages/Articles.jsx
@@ -5,9 +5,7 @@ import {
   deleteDoc,
   doc,
   updateDoc,
-  arrayUnion,
-  arrayRemove,
-  onSnapshot,
+  getDocs,
 } from "firebase/firestore";
 import { format } from "date-fns";
 import { ToastContainer, toast } from "react-toastify";
@@ -18,25 +16,26 @@ import "jodit";
 import JoditEditor from "jodit-react";
 import { NavLink } from "react-router-dom";
 import { BiLike, BiSolidLike } from "react-icons/bi";
+import { useQuery, useQueryClient } from "react-query";
 
 export const Articles = () => {
+  const queryClient = useQueryClient();
   const editorRef = useRef(null);
   const { user, firestore } = useAuthContext();
-  const [articles, setArticles] = useState([]);
   const [editArticleId, setEditArticleId] = useState(null);
   const [editTitle, setEditTitle] = useState("");
   const [editContent, setEditContent] = useState("");
   const [searchQuery, setSearchQuery] = useState("");
   const [sortOrder, setSortOrder] = useState("newest");
   const [selectedCategory, setSelectedCategory] = useState("all");
-  const [loading, setLoading] = useState(false);
-  const config = useMemo(() => {
-    return {
+  const config = useMemo(
+    () => ({
       readonly: false,
       height: "500px",
       uploader: { insertImageAsBase64URI: true },
-    };
-  }, []);
+    }),
+    []
+  );
   const [categories] = useState([
     "Proqramlaşdırma",
     "Qrafik dizayn",
@@ -48,60 +47,34 @@ export const Articles = () => {
   const isAdmin = user && user.email === "alyvdev@gmail.com";
   const articleContentRef = useRef(null);
 
-  const fetchArticlesFromFirestore = () => {
-    setLoading(true);
+  // React Query ile veri çekme ve önbelleğe alma
+  const { data: articles, isLoading } = useQuery(["articles"], async () => {
+    const querySnapshot = await getDocs(collection(firestore, "articles"));
+    return querySnapshot.docs.map((doc) => ({
+      id: doc.id,
+      ...doc.data(),
+    }));
+  });
 
-    // LocalStorage'den verileri kontrol edip, makaleleri ID bazlı alıyoruz
-    const localArticles = [];
-    const localStorageKeys = Object.keys(localStorage);
-
-    localStorageKeys.forEach((key) => {
-      if (key.startsWith("article_")) {
-        const storedArticle = JSON.parse(localStorage.getItem(key));
-        localArticles.push(storedArticle);
-      }
-    });
-
-    if (localArticles.length > 0) {
-      setArticles(localArticles);
-      setLoading(false);
-    }
-
-    // Firebase'den gerçek zamanlı verileri alıyoruz
-    const articlesRef = collection(firestore, "articles");
-    const unsubscribe = onSnapshot(articlesRef, (querySnapshot) => {
-      const articlesList = querySnapshot.docs.map((doc) => ({
-        id: doc.id,
-        ...doc.data(),
-      }));
-
-      // Gelen her makale için localStorage'a kaydediyoruz
-      articlesList.forEach((article) => {
-        localStorage.setItem(`article_${article.id}`, JSON.stringify(article));
-      });
-
-      setArticles(articlesList);
-      setLoading(false);
-    });
-
-    return () => unsubscribe(); // Temizlik işlemi
-  };
-
+  // Makale silme
   const deleteArticle = async (id) => {
+    if (!articles) return; // Ensure articles is defined
     const articleToDelete = articles.find((article) => article.id === id);
-    if (articleToDelete.userId === user.uid || isAdmin) {
+    if (articleToDelete && (articleToDelete.userId === user.uid || isAdmin)) {
       try {
         await deleteDoc(doc(firestore, "articles", id));
+        localStorage.removeItem(`article_${id}`);
         toast.success("Məqalə silindi!");
-        fetchArticlesFromFirestore(); // Veri silindiği için yeniden yükle
       } catch (error) {
         toast.error("Məqalə silinə bilmədi!");
       }
+      queryClient.invalidateQueries("articles");
     } else {
       toast.error("Bu məqaləyi silməyə icazəniz yoxdur!");
     }
   };
 
+  // Düzenleme durumunu yönetme
   const handleEditChange = (e, field) => {
     if (field === "title") {
       setEditTitle(e.target.value);
@@ -110,178 +83,205 @@ export const Articles = () => {
     }
   };
 
+  // Makale güncelleme
   const updateArticle = async (id) => {
     try {
       const articleRef = doc(firestore, "articles", id);
-
-      // Access the editor's value using the ref
       const updatedContent = editorRef.current.value;
 
+      // Update Firestore
       await updateDoc(articleRef, {
         title: editTitle,
-        content: updatedContent, // Set the updated content from the editor
+        content: updatedContent,
         updatedAt: new Date(),
       });
+
+      // Update local cache
+      queryClient.setQueryData(["articles"], (oldArticles) =>
+        oldArticles.map((article) =>
+          article.id === id
+            ? { ...article, title: editTitle, content: updatedContent }
+            : article
+        )
+      );
+
       toast.success("Məqalə güncəlləndi!");
       setEditArticleId(null);
-      fetchArticlesFromFirestore(); // Fetch updated articles from Firestore
     } catch (error) {
-      toast.error("Məqalə güncəllənə bilmədi!");
+      console.error("Güncellemeye hata:", error);
+      toast.error("Məqalə güncəllənə bilmədi: " + error.message);
     }
   };
 
+  // Beğenme işlemi
   const toggleLike = async (articleId) => {
+    const article = articles.find((article) => article.id === articleId);
+    const updatedLikes = article.likes.includes(user.uid)
+      ? article.likes.filter((uid) => uid !== user.uid)
+      : [...article.likes, user.uid];
+
     try {
       const articleRef = doc(firestore, "articles", articleId);
-      const article = articles.find((article) => article.id === articleId);
+      // Update Firestore
+      await updateDoc(articleRef, {
+        likes: updatedLikes,
+      });
 
-      if (article.likes.includes(user.uid)) {
-        await updateDoc(articleRef, {
-          likes: arrayRemove(user.uid),
-        });
-      } else {
-        await updateDoc(articleRef, {
-          likes: arrayUnion(user.uid),
-        });
-      }
-
-      fetchArticlesFromFirestore();
+      // Update local cache
+      queryClient.setQueryData(["articles"], (oldArticles) =>
+        oldArticles.map((article) =>
+          article.id === articleId
+            ? { ...article, likes: updatedLikes }
+            : article
+        )
+      );
     } catch (error) {
       console.error("Bəyənmək baş tutmadı:", error);
       toast.error("Bəyənmək baş tutmadı!");
     }
   };
 
-  useEffect(() => {
-    if (user) {
-      const unsubscribe = fetchArticlesFromFirestore(); // Verileri gerçek zamanlı dinle
-      return unsubscribe; // Temizlik işlemi (unsubscribe)
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [user, firestore]);
-
+  // Kod vurgulama
   useEffect(() => {
     if (articleContentRef.current) {
       hljs.highlightAll();
     }
   }, [articles]);
 
-  // Filter and sort the articles
-  const filteredAndSortedArticles = articles
-    .filter((article) => {
-      const matchesSearch = article.title
-        .toLowerCase()
-        .includes(searchQuery.toLowerCase());
-      const matchesCategory =
-        selectedCategory === "all" || article.category === selectedCategory;
-      return matchesSearch && matchesCategory;
-    })
-    .sort((a, b) => {
-      if (sortOrder === "newest") {
-        return b.createdAt.seconds - a.createdAt.seconds;
-      } else if (sortOrder === "oldest") {
-        return a.createdAt.seconds - b.createdAt.seconds;
-      } else if (sortOrder === "A-Z") {
-        return a.title.localeCompare(b.title);
-      } else {
-        return b.title.localeCompare(a.title);
-      }
-    });
+  // Makaleleri filtreleme ve sıralama
+  const filteredAndSortedArticles = useMemo(() => {
+    return articles
+      ?.filter((article) => {
+        const matchesSearch = article.title
+          .toLowerCase()
+          .includes(searchQuery.toLowerCase());
+        const matchesCategory =
+          selectedCategory === "all" || article.category === selectedCategory;
+        return matchesSearch && matchesCategory;
+      })
+      .sort((a, b) => {
+        if (sortOrder === "newest") {
+          return b.createdAt.seconds - a.createdAt.seconds;
+        } else if (sortOrder === "oldest") {
+          return a.createdAt.seconds - b.createdAt.seconds;
+        } else if (sortOrder === "A-Z") {
+          return a.title.localeCompare(b.title);
+        } else {
+          return b.title.localeCompare(a.title);
+        }
+      });
+  }, [articles, searchQuery, selectedCategory, sortOrder]);
 
   return (
-    <section className=" p-5 flex flex-col  min-h-screen overflow-y-auto">
-      {/* Toast bildirimlerini göstermek için ToastContainer */}
+    <section className="p-5 flex flex-col overflow-y-auto w-full">
       <ToastContainer position="top-center" autoClose={2000} />
 
-      {/* Yüklenme durumu */}
-      {loading && <div className="text-center text-gray-500">Yüklənir...</div>}
-
-      {/* Arama ve filtreleme seçeneklerini içeren bölüm */}
-      <div className="flex sm:flex-col lg:flex-row-reverse sm:items-start lg:items-center lg:justify-between gap-3 mb-4">
-        {/* Arama kutusu */}
-        <div className="relative">
-          <div className="absolute inset-y-0 start-0 flex items-center ps-3 pointer-events-none">
+      {isLoading && user && (
+        <div className="w-full h-screen flex items-center justify-center">
+          <div className="text-center">
             <svg
-              className="w-4 h-4 text-gray-500 dark:text-gray-400"
-              aria-hidden="true"
+              className="animate-spin h-10 w-10 text-blue-500 mx-auto mb-4"
               xmlns="http://www.w3.org/2000/svg"
               fill="none"
-              viewBox="0 0 20 20"
+              viewBox="0 0 24 24"
             >
-              <path
+              <circle
+                className="opacity-25"
+                cx="12"
+                cy="12"
+                r="10"
                 stroke="currentColor"
-                strokeLinecap="round"
-                strokeLinejoin="round"
-                strokeWidth="2"
-                d="m19 19-4-4m0-7A7 7 0 1 1 1 8a7 7 0 0 1 14 0Z"
-              />
+                strokeWidth="4"
+              ></circle>
+              <path
+                className="opacity-75"
+                fill="currentColor"
+                d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z"
+              ></path>
             </svg>
+            <p className="text-lg text-gray-700">Məqalələr yüklənir...</p>
           </div>
-          <input
-            type="search"
-            id="default-search"
-            className="block w-full p-1 ps-10 text-sm  focus:outline-none border border-gray-500 rounded-lg "
-            placeholder="Axtarışa başlayın..."
-            value={searchQuery}
-            onChange={(e) => setSearchQuery(e.target.value)}
-          />
         </div>
+      )}
+      {user && (
+        <div className="flex sm:flex-col lg:flex-row-reverse sm:items-start lg:items-center lg:justify-between gap-3 mb-4">
+          <div className="relative">
+            <div className="absolute inset-y-0 start-0 flex items-center ps-3 pointer-events-none">
+              <svg
+                className="w-4 h-4 text-gray-500 dark:text-gray-400"
+                aria-hidden="true"
+                xmlns="http://www.w3.org/2000/svg"
+                fill="none"
+                viewBox="0 0 20 20"
+              >
+                <path
+                  stroke="currentColor"
+                  strokeLinecap="round"
+                  strokeLinejoin="round"
+                  strokeWidth="2"
+                  d="m19 19-4-4m0-7A7 7 0 1 1 1 8a7 7 0 0 1 14 0Z"
+                />
+              </svg>
+            </div>
+            <input
+              type="search"
+              id="default-search"
+              className="block w-full p-1 ps-10 text-sm  focus:outline-none border border-gray-500 rounded-lg "
+              placeholder="Axtarışa başlayın..."
+              value={searchQuery}
+              onChange={(e) => setSearchQuery(e.target.value)}
+            />
+          </div>
 
-        {/* Sıralama ve kategori seçimi */}
-        <div className="w-auto flex">
-          <label htmlFor="sortOrder" className=""></label>
-          <select
-            id="sortOrder"
-            value={sortOrder}
-            onChange={(e) => setSortOrder(e.target.value)}
-            className="p-1 mr-2 border border-gray-300 rounded-lg focus:outline-none"
-          >
-            <option value="newest">Yenilər</option>
-            <option value="oldest">Köhnələr</option>
-            <option value="A-Z">A-Z</option>
-            <option value="Z-A">Z-A</option>
-          </select>
+          <div className="w-auto flex">
+            <label htmlFor="sortOrder" className=""></label>
+            <select
+              id="sortOrder"
+              value={sortOrder}
+              onChange={(e) => setSortOrder(e.target.value)}
+              className="p-1 mr-2 border border-gray-300 rounded-lg focus:outline-none"
+            >
+              <option value="newest">Yenilər</option>
+              <option value="oldest">Köhnələr</option>
+              <option value="A-Z">A-Z</option>
+              <option value="Z-A">Z-A</option>
+            </select>
 
-          <label htmlFor="category" className=""></label>
-          <select
-            id="category"
-            value={selectedCategory}
-            onChange={(e) => setSelectedCategory(e.target.value)}
-            className="p-1 border border-gray-300 rounded-lg focus:outline-none"
-          >
-            <option value="all">Hamısı</option>
-            {categories.map((category) => (
-              <option key={category} value={category}>
-                {category}
-              </option>
-            ))}
-          </select>
+            <label htmlFor="category" className=""></label>
+            <select
+              id="category"
+              value={selectedCategory}
+              onChange={(e) => setSelectedCategory(e.target.value)}
+              className="p-1 border border-gray-300 rounded-lg focus:outline-none"
+            >
+              <option value="all">Hamısı</option>
+              {categories.map((category) => (
+                <option key={category} value={category}>
+                  {category}
+                </option>
+              ))}
+            </select>
+          </div>
         </div>
-      </div>
+      )}
 
-      {/* Makale kartlarının listelendiği bölüm */}
       <div className="gap-2 flex flex-col">
-        {filteredAndSortedArticles.length > 0 ? (
+        {filteredAndSortedArticles?.length > 0 ? (
           filteredAndSortedArticles.map((article) => (
             <div
               key={article.id}
-              className={`p-2 border-2 border-gray-200 rounded-xl relative w-full${
+              className={`p-2  bg-white shadow-lg rounded-xl relative w-full${
                 editArticleId === article.id ? "bg-gray-100" : ""
               }`}
             >
-              {/* Makale başlığı */}
               <h3 className="text-lg sm:mr-5 lg:mr-0 font-semibold cursor-pointer hover:text-blue-500">
                 <NavLink to={`/articles/${article.id}`}>
                   {article.title}
                 </NavLink>
               </h3>
-
-              {/* Makale kategorisi */}
               <span className="text-sm text-blue-900">
                 Kateqoriya: {article.category}
               </span>
-
-              {/* Makale yazıldığı tarih */}
               <p className="text-xs text-gray-500 mb-1">
                 Yazıldığı tarix:{" "}
                 {format(
@@ -290,7 +290,6 @@ export const Articles = () => {
                 )}
               </p>
 
-              {/* Beğen butonu */}
               <button
                 onClick={(e) => {
                   e.stopPropagation();
@@ -306,7 +305,6 @@ export const Articles = () => {
                 <span className="ml-1">{article.likes.length}</span>
               </button>
 
-              {/* Makale düzenleme ve silme butonları (kullanıcı yetkiliyse gösterilir) */}
               {(article.userId === user.uid || isAdmin) && (
                 <div className="w-full flex justify-end gap-2">
                   {editArticleId !== article.id && (
@@ -334,7 +332,6 @@ export const Articles = () => {
                     </>
                   )}
 
-                  {/* Makale düzenleme formu */}
                   {editArticleId === article.id && (
                     <div className="w-full">
                       <input
@@ -347,10 +344,7 @@ export const Articles = () => {
                         ref={editorRef}
                         value={editContent}
                         config={config}
-                        onBlur={(newContent) => handleEditChange(newContent)}
-                        onChange={(newContent) => {
-                          handleEditChange(newContent);
-                        }}
+                        onChange={(newContent) => handleEditChange(newContent)}
                         className="w-full"
                       />
 
@@ -379,11 +373,15 @@ export const Articles = () => {
             </div>
           ))
         ) : user ? (
-          // Kullanıcı giriş yapmışsa ancak makale yoksa
           <p>Məqalə yoxdur.</p>
         ) : (
-          // Kullanıcı giriş yapmamışsa
-          <p>Məqalələri görmək üçün zəhmət olmasa giriş edin.</p>
+          <div className="h-[85vh] flex items-center justify-center">
+            <div className="bg-white p-8 rounded-lg shadow-xl text-center max-w-lg">
+              <p className="text-xl font-semibold text-black mb-4">
+                Məqalələri oxumaq üçün daxil olun
+              </p>
+            </div>
+          </div>
         )}
       </div>
     </section>
